added more icons and changed likes
