+++ src/components/CommentSection/CommentSectionContainer.js
@@ -4,35 +4,11 @@ import CommentInput from "./CommentInput";
 import Comment from "./Comment";
 import "./Comment.css";
 import TimeStamp from "../PostsContainer/TimeStamp";
+import DateConverter from "./DateConverter"
 
 const CommentSection = props => {
   const [comments, setComments] = useState(props.comments)
-  const [timeSince, setTimeSince] = useState(()=>{
-    let dateValues = props.time.split(" ")
-    let todayDate = Date(Date.now()).split(" ");
-    console.log(Date.parse(`July 1, 2020 ${todayDate[4]}`))
-    console.log(dateValues, todayDate)
-    if(todayDate[3] !== dateValues[2]){
-      return `${parseInt(todayDate[3]) - parseInt(dateValues[2])} years ago`
-    }
-    if(todayDate[1] !== dateValues[0]){
-      let postMonth = new Date(Date.parse(`${dateValues[0]} 1, 2020`)).getMonth()+1
-      let todayMonth = new Date(Date.parse(`${todayDate[1]} 1, 2020`)).getMonth()+1
-      return `${todayMonth - postMonth} months ago`
-    }
-    let postDay = parseInt(todayDate(dateValues[1]))
-    let todayDay = parseInt(todayDate[2]);
-    if(postDay !== todayDay){
-      return `${todayDay - postDay} days ago`
-    }
-    let postHours = Date.parse(`July 1, 2020 ${dateValues[3]}`)
-    if(dateValues[4] === "pm" && parseInt(dateValues[3]) !== 12){
-      postHours += 3600000;
-    }
-    else if(dateValues[4] === "am" && parseInt(dateValues[3]) === 12){
-      postHours -= 3600000;
-    }
-  })
+  const [timeSince, setTimeSince] = useState(DateConverter(props.time))
   const [newComment, setnewComment] = useState("")
 
   function submitComment(e){
